Fix SQL Munging on CSV Field Queries
  - fixed a class of bugs where the description and tag fields could return the wrong rows
