+++ kamailio/configs/kamailio.cfg
@@ -2508,7 +2508,7 @@ route[SET_CALLINFO] {
 		if (sql_xquery("kam", "select rpid as gwgroupid from subscriber where username='$au'", "rows") == 1) {
 			$avp(src_gwtype) = (str)FLT_PBX;
 			$avp(src_gwgroupid) = $xavp(rows=>gwgroupid);
-			if (sql_xquery("kam", "select gwid from dr_gateways where address='$var(received_addr)' AND description LIKE '%gwgroup:$avp(src_gwgroupid)%'", "rows") == 1) {
+			if (sql_xquery("kam", "select gwid from dr_gateways where address='$var(received_addr)' AND description REGEXP 'gwgroup:$avp(src_gwgroupid)(,|$$)'", "rows") == 1) {
 				$avp(src_gwid) = $xavp(rows=>gwid);
 			}
 		}
@@ -3479,7 +3479,7 @@ event_route[usrloc:contact-expired] {
 		$var(src_gwgroupid) = $xavp(rows=>gwgroupid);
 		$var(received_addr) = $(ulc(exp=>received){re.subst,/^(sip:|sips:)?(.*)$/\2/});
 
-		if (sql_xquery("kam", "SELECT gwid FROM dr_gateways WHERE address='$var(received_addr)' AND description LIKE '%gwgroup:$var(src_gwgroupid)%'", "rows") == 1) {
+		if (sql_xquery("kam", "SELECT gwid FROM dr_gateways WHERE address='$var(received_addr)' AND description REGEXP 'gwgroup:$var(src_gwgroupid)(,|$$)'", "rows") == 1) {
 			$var(src_gwid) = $xavp(rows=>gwid);
 
 			xlog("L_DBG", "found gateway $var(src_gwid) for address $var(received_addr), removing from gwgroup $var(src_gwgroupid)\n");
